НЕДИЗАЙН, реализована автоматическая подстановка баллов, выбор для 3 параметров карточки

+++ lib/cubit/card_cubit/card_cubit.dart
@@ -32,7 +32,7 @@ class CardCubit extends Cubit<CardCubitState> {
     required String achievementStatus,
     required String achievementLevel,
     required String documentProof,
-    required int points,
+    required double points,
   }) async {
     try {
       final newEvent = EventModel(
